BankToEnote/requirements.txt
python-dotenv==1.0.1
mysql-connector-python==9.0.0
requests==2.32.3
